+++ packages/keyring-api/src/api/v2/create-account/index.ts
@@ -92,3 +92,46 @@ export const CreateAccountOptionsStruct = selectiveUnion((value: any) => {
  * Represents the available options for creating a new account.
  */
 export type CreateAccountOptions = Infer<typeof CreateAccountOptionsStruct>;
+
+/**
+ * Asserts that a given create account option type is supported by the keyring.
+ *
+ * @example
+ * ```ts
+ * createAccounts(options: CreateAccountOptions) {
+ *   assertCreateAccountOptionIsSupported(options, [
+ *     ${AccountCreationType.Bip44DeriveIndex},
+ *     ${AccountCreationType.Bip44DeriveIndexRange},
+ *   ] as const);
+ *
+ *   // At this point, TypeScript knows that options.type is either Bip44DeriveIndex or Bip44DeriveIndexRange.
+ *   if (options.type === AccountCreationType.Bip44DeriveIndex) {
+ *     ... // Handle Bip44DeriveIndex case.
+ *   } else {
+ *     ... // Handle Bip44DeriveIndexRange case.
+ *   }
+ *   ...
+ *   return accounts;
+ * }
+ * ```
+ *
+ * @param options - The create account option object to check.
+ * @param supportedTypes - The list of supported create account option types for this keyring.
+ * @throws Will throw an error if the provided options are not supported.
+ */
+export function assertCreateAccountOptionIsSupported<
+  Options extends CreateAccountOptions,
+  // We use template literal types to enforce string-literal over strict enum values.
+  Type extends `${CreateAccountOptions['type']}`,
+>(
+  options: Options,
+  supportedTypes: readonly `${Type}`[],
+  // Use intersection to avoid widening `type` beyond `Options['type']`.
+): asserts options is Options & { type: `${Type}` & `${Options['type']}` } {
+  const { type } = options;
+  const types: readonly CreateAccountOptions['type'][] = supportedTypes;
+
+  if (!types.includes(type)) {
+    throw new Error(`Unsupported create account option type: ${type}`);
+  }
+}